// src/enums/Errors.enum.ts
export enum Errors {
    EMPTY_FIELD = 'This field has to be filled.',
    EMAIL_WRONG_FORMAT = 'Email has wrong format.',
    EASY_PASSWORD_WRONG_FORMAT = 'Password has to contains from 8 to 100 symbols, excluding special characters ',
    PASSWORDS_DOES_NOT_MATCH = 'Passwords does not matches.',
    PHONE_NUMBER_WRONG_FORMAT = 'Phone number has wrong format.',
    POSTAL_CODE_WRONG_FORMAT = 'Postal code has wrong format.',
    ONLY_LETTERS = 'Field has to contain only letters.',
    SPECIAL_CHARACTERS_NOT_ALLOWED = 'Field cannot contains any special characters.',
}